+++ paper/paper.bib
@@ -96,7 +96,8 @@
   series        = {OSA Technical Digest},
   pages         = {DTh2A.2},
   publisher     = {Optical Society of America},
-  url           = {http://www.osapublishing.org/abstract.cfm?URI=BODA-2021-DTh2A.2}
+  doi           = {10.1364/boda.2021.dth2a.2},
+  url           = {https://doi.org/10.1364/boda.2021.dth2a.2}
 }
 
 @Article{Cui2021OpticsExpress,
@@ -160,6 +161,7 @@
   volume    = {9148},
   publisher = {SPIE},
   type      = {Conference Proceedings},
+  doi       = {10.1117/12.2054470},
   url       = {https://doi.org/10.1117/12.2054470},
 }
 
@@ -196,6 +198,7 @@
   pages   = {13-62},
   year    = {1993},
   type    = {Journal Article},
+  doi     = {10.1146/annurev.aa.31.090193.000305},
   url     = {https://www.annualreviews.org/doi/pdf/10.1146/annurev.aa.31.090193.000305},
 }
 
