### 2022-01-11 #4023 @abhilash12iec002
Yes, I can review the said paper.

Best,

Abhilash Singh
PhD Scholar, 1810302
Fluvial Geomorphology and Remote Sensing Lab,
Department of EES, AB-2, Lab-10,
Indian Institute of Science Education and Research Bhopal
mobile: +91-9990674147
***@***.***
www.abhilashsingh.net
***@***.***>

On Tue, Jan 11, 2022 at 1:29 PM Jonatan Alvelid ***@***.***>
wrote:

> Hi @diehlpk <https://github.com/diehlpk>, thank you for considering me,
> but I am unfortunately not available to review currently.
> I hope I will be considered again in the future and will happily review
> then, and thanks for your understanding.
>
> Best,
> Jonatan
>
