### 2022-01-20 #4023 @jiahecui
> @jiahecui Please find my comments below; I have read the complete article, “SenAOReFoc: A Closed-Loop Sensorbased Adaptive Optics and Remote Focusing Control Software” with full consideration. The title is short and sufficient. The summary part is nicely written in a very concise way. In the statement of need section, there exists a scope of improvement. For example; “The functionality of the software has also been tested on different operating systems (Windows/macOS/Linux) for sake of generality.” For this line, a concluding remark on the performance on each platform must be added. Also, the lines from 71 to 73 must be explained in detail. I think these two points are very crucial points and a good set of motivation for this software. I am very much satisfied with the Example usage section. Apart from reading the article, I have spent a good chunk of time using the SenAOReFoc package in the due course of this review. I am satisfied with the functionality of this software.
>
> Overall, the submitted manuscript is well written, and ably demonstrates the community requirement for this open-source tool. I recommend publication after incorporating the suggestions discussed above.

@abhilash12iec002 Thank you very much for the thorough review and detailed comments. Your time is greatly appreciated! We will make the revisions as soon as possible!
