### 2022-02-24 #4075 @diehlpk
Hi @jiahecui,

I started to look al the paper from the editorial perspective and have the following comments
`
* It seems the citations are not working

> If you're performing remote focusing using the calibration procedure set out in this software, please also cite either [@Cui2021Biophotonics] or [@Cui2021OpticsExpress].`

* Do you have any tool, like pydoc, to generate the documentation for your code?

* Dou you have any code coverage test?
