### 2022-02-24 #4075 @jiahecui
Hi @diehlp, thank you very much for the message!

> It seems the citations are not working
* I use whedon https://whedon.theoj.org/ to generate the pdf, the citations did not cause a problem previously and I think they are ok now also? The pdf has been generated fine I think?

> Do you have any tool, like pydoc, to generate the documentation for your code?
* No unfortunately we have never used pydoc. Is this required? We have prepared a very detailed README.md file in the root repo that explains the functionalities of the code.

> Do you have any code coverage test?
* Do you mean coverage as in operating platforms or equipment range? We have stated in the paper that we tested the software on Linux and macOS. These were tested using the dummy mode as there is no option for us to test hardware on these platforms. One of the reviewers asked for some concluding remarks on the results. I am still waiting for input from my colleague as he has been ill for a long time and is just resuming work. Please give us a few more days and we will provide updates as soon as possible. Regarding equipment range, we stated at the beginning that this software targets towards control of Shack-Hartmann wavefront sensors and deformable mirrors. Please let us know if this answers your question.

> Do you have some automated testing or does the user needs to call all the listed scripts per hand?
* The user will need to press the specified GUI buttons in the order as stated in the README.md to properly test the software and unfortunately there is no way for automation. This is a device control software and one of the critical functionalities is that the user can control the position of the search block area using the left/right/up/down arrows on their keyboard during experiments. This is an essential step that cannot be automated. But there aren't many buttons to press for software testing, and they are clearly stated in the README.md, so shouldn't be a problem.

Thank you very much again and please let us know if there are further questions. We hope to respond to the first reviewer's comments very soon.
