### 2022-02-24 #4075 @diehlpk
> > It seems the citations are not working
>
>     * I use whedon https://whedon.theoj.org/ to generate the pdf, the citations did not cause a problem previously and I think they are ok now also? The pdf has been generated fine I think?

@jiahecui I was talking about the Readme.md of the GitHub repo. Sorry for the confusion.
