### 2022-02-24 #4075 @diehlpk
> > Do you have some automated testing or does the user needs to call all the listed scripts per hand?
>
>     * The user will need to press the specified GUI buttons in the order as stated in the README.md to properly test the software and unfortunately there is no way for automation. This is a device control software and one of the critical functionalities is that the user can control the position of the search block area using the left/right/up/down arrows on their keyboard during experiments. This is an essential step that cannot be automated. But there aren't many buttons to press for software testing, and they are clearly stated in the README.md, so shouldn't be a problem.

In that case, it is fine.
