### 2022-02-24 #4075 @diehlpk
> > Do you have any code coverage test?
>
>     * Do you mean coverage as in operating platforms or equipment range? We have stated in the paper that we tested the software on Linux and macOS. These were tested using the dummy mode as there is no option for us to test hardware on these platforms. One of the reviewers asked for some concluding remarks on the results. I am still waiting for input from my colleague as he has been ill for a long time and is just resuming work. Please give us a few more days and we will provide updates as soon as possible. Regarding equipment range, we stated at the beginning that this software targets towards control of Shack-Hartmann wavefront sensors and deformable mirrors. Please let us know if this answers your question.

I was more thinking about like that

{Coverage.py](https://coverage.readthedocs.io/en/6.3.2/#coverage-py)

Coverage.py is a tool for measuring code coverage of Python programs. It monitors your program, noting which parts of the code have been executed, then analyzes the source to identify code that could have been executed but was not.

If the tests are executed, how many lines of code were reached and tested?
