### 2022-02-27 #4075 @jiahecui
Dear Patrick,

Hello, and many thanks for the messages. I don’t know why but I have not been able to access any GitHub websites since Friday, including this review thread, I can only reply by email for now. Chrome gives me the following error code NET::ERR_CERT_COMMON_NAME_INVALID. It seems to come from an imposed security update from our Dept. I will need to contact our IT to find out.

It the meantime, I have installed coverage.py, and ran the test coverage of the main app.py script. The results are attached as screenshot. This only represents the coverage for executing the dummy test steps specified in README.md. More buttons can be pressed to leverage a higher coverage, but they are not necessary, as without hardware feedback, most functions are only to test the execution of the scripts and do not reflect algorithmic control. Even if hardware were connected, as different functions are required for different control purposes, the coverage would be different. Hope this helps and please let me know if there is any further information I can provide. Many thanks again.

Best wishes,
Jiahe

发件人: Patrick Diehl ***@***.***>
发送时间: 2022年2月24日 19:32
收件人: openjournals/joss-reviews ***@***.***>
抄送: Jiahe ***@***.***>; Mention ***@***.***>
主题: Re: [openjournals/joss-reviews] [REVIEW]: SenAOReFoc: A Closed-Loop Sensorbased Adaptive Optics and Remote Focusing Control Software (Issue #4075)

Do you have any code coverage test?

* Do you mean coverage as in operating platforms or equipment range? We have stated in the paper that we tested the software on Linux and macOS. These were tested using the dummy mode as there is no option for us to test hardware on these platforms. One of the reviewers asked for some concluding remarks on the results. I am still waiting for input from my colleague as he has been ill for a long time and is just resuming work. Please give us a few more days and we will provide updates as soon as possible. Regarding equipment range, we stated at the beginning that this software targets towards control of Shack-Hartmann wavefront sensors and deformable mirrors. Please let us know if this answers your question.

I was more thinking about like that

{Coverage.py](https://coverage.readthedocs.io/en/6.3.2/#coverage-py)

Coverage.py is a tool for measuring code coverage of Python programs. It monitors your program, noting which parts of the code have been executed, then analyzes the source to identify code that could have been executed but was not.

If the tests are executed, how many lines of code were reached and tested?
