### 2022-02-28 #4075 @diehlpk
> Hi @diehlp, thank you very much for the messages again!
>
> > I was talking about the Readme.md of the GitHub repo. Sorry for the confusion.
> > No worries at all. I have updated this in README.md and it directly gives links to the references now.
>
> > I was more thinking about like that {Coverage.py](https://coverage.readthedocs.io/en/6.3.2/#coverage-py). If the tests are executed, how many lines of code were reached and tested?
> > Thanks for recommending this function.  I have installed coverage.py, and ran the test coverage of the main app.py script. The results are attached as a screenshot. This only represents the coverage for executing the dummy test steps specified in README.md. More buttons can be pressed to leverage a higher coverage, but they are not necessary, as without hardware feedback, most functions are only to test the execution of the scripts and do not reflect algorithmic control. Even if hardware were connected, as different functions are required for different control purposes, the coverage would be different. Hope this helps and please let me know if there is any further information I can provide. Many thanks again! <img alt="Coverage" width="661" src="https://user-images.githubusercontent.com/44090103/155975130-d2dff6b0-7001-46a8-a738-1394fd9f04e5.PNG">

@jiahecui Could you add some documentation how to run the tests this way?
