### 2022-02-28 f5fa72a
Coverage test added to README.md
