### 2022-02-28 ce0c493
Coverage test added to README.md
