### 2022-02-28 1898cdd
Coverage test added to README.md
