### 2022-02-28 #4075 @jiahecui
Hi @diehlp, this is a very helpful suggestion. I have now included this information (including how to run the coverage test in the command prompt + result screenshot) in the `Software functionality tests` section of README.md immediately after standard testing instructions. Thanks again.
