### 2022-03-01 #4075 @jiahecui
> Thanks that looks good. I will read the paper tomorrow and let you know my final comments.

@diehlp That's great, thank you so much!

> We hope to respond to the first reviewer's comments very soon.
> @jiahecui please let me know once you have done this as well.

@diehlp The second comment of the reviewer has now been addressed in the paper. Lines 73-77 now provide explanations to lines 71-73. I am waiting for a concluding remark from my colleague for the first comment on different platforms. I will update you about this as soon as possible. Many thanks again!
