### 2022-03-02 #4075 @diehlpk
@jiahecui My only remark left is for the documentation. For example, you documented the classes like here

https://github.com/jiahecui/SenAOReFoc/blob/1898cdd048b4149b73440969b0f976805774d99a/sensorbasedAO/AO_slopes.py#L18

which is good. However, some class members are not documented like here

https://github.com/jiahecui/SenAOReFoc/blob/1898cdd048b4149b73440969b0f976805774d99a/sensorbasedAO/AO_slopes.py#L20

I think it would not be much effort to document these to make the code easier to understand.
