### 2022-03-02 #4075 @diehlpk
@jiahecui

These are the last two items before I can recommend the paper for acceptance

- [ ]  Documentation of class members
- [ ] Reviwer's 1 comment
