### 2022-03-02 c318690
Documentation added for Signal classes in all scripts
