+++ sensorbasedAO/conversion.py
@@ -12,11 +12,22 @@ class Conversion(QObject):
     """
     Generates slope - zernike conversion matrix and zernike matrix for a given pupil shape
     """
+    # Signal class for starting an event
     start = Signal()
+
+    # Signal class for writing DM parameters into HDF5 file
     write = Signal()
+
+    # Signal class for exiting slope - zernike conversion event
     done = Signal()
+
+    # Signal class for emitting a message in the message box
     message = Signal(object)
+
+    # Signal class for raising an error
     error = Signal(object)
+
+    # Signal class for updating DM control parameters
     info = Signal(object)
 
     def __init__(self, settings, debug = False):
