+++ sensorbasedAO/centroiding.py
@@ -16,14 +16,31 @@ class Centroiding(QObject):
     """
     Performs system aberration calibration
     """
+    # Signal class for starting an event
     start = Signal()
+
+    # Signal class for writing system calibration information into HDF5 file
     write = Signal()
+
+    # Signal class for writing search block information into HDF5 file
     SB_write = Signal()
+
+    # Signal class for exiting system calibration event
     done = Signal()
+
+    # Signal class for raising an error
     error = Signal(object)
+
+    # Signal class for displaying a SH spot image
     image = Signal(object)
+
+    # Signal class for emitting a message in the message box
     message = Signal(object)
+
+    # Signal class for updating system calibration parameters
     info = Signal(object)
+
+    # Signal class for updating search block information
     SB_info = Signal(object)
     
     def __init__(self, sensor, mirror, settings, debug = False):
