+++ sensorbasedAO/calibration_RF.py
@@ -19,14 +19,31 @@ class Calibration_RF(QObject):
     """
     Calibrates remote focusing using nulling correction 
     """
+    # Signal class for starting an event
     start = Signal()
+
+    # Signal class for writing DM parameters into HDF5 file
     write = Signal()
+
+    # Signal class for writing RF calibration parameters into HDF5 file
     calib_write = Signal()
+
+    # Signal class for exiting RF calibration event
     done = Signal()
+
+    # Signal class for raising an error
     error = Signal(object)
+
+    # Signal class for displaying a SH spot image
     image = Signal(object)
+
+    # Signal class for emitting a message in the message box
     message = Signal(object)
+
+    # Signal class for updating DM control parameters
     info = Signal(object)
+
+    # Signal class for updating RF control parameters
     calib_info = Signal(object)
 
     def __init__(self, sensor, mirror, settings, debug = False):
