### 2022-03-02 #4075 @jiahecui
@diehlp Thank you very much for reading through the code and for suggesting the last two items before acceptance.

> My only remark left is for the documentation. I think it would not be much effort to document these to make the code easier to understand.

This suggestion is very helpful thanks. I have now added documentation for Signal classes in all scripts, in the following way for example in AO_slopes.py.

```
# Signal class for starting an event
    start = Signal()

    # Signal class for writing slope AO information into HDF5 file
    write = Signal()

    # Signal class for exiting slope AO event
    done = Signal(object)

    # Signal class for exiting remote focusing event
    done2 = Signal(object)

    # Signal class for raising an error
    error = Signal(object)

    # Signal class for displaying a SH spot image
    image = Signal(object)

    # Signal class for emitting a message in the message box
    message = Signal(object)

    # Signal class for updating slope AO information
    info = Signal(object)
```

Please let me know if this looks good or if more information should be provided. Thank you very much again. The last item I am still chasing my colleague.
