### 2022-03-04 a6e6195
Platform testing conclusions added
