### 2022-03-04 1a05eea
Platform testing conclusions added
