+++ paper/paper.md
@@ -113,7 +113,7 @@ microscopes with a simple and user-friendly architecture. The functionality of t
 on different operating systems (Windows/Linux/macOS) for sake of generality. The software fully operates in 
 Windows under both 'debug mode' and 'standard mode' for hardware control. For Linux and macOS, functionality tests
 in 'debug mode' passes for both systems when hardware packages are not imported. Python-bindings for Ximea cameras
-and Python 3.8-binding for Alpao DMs are compatible with Linux systems, and other Python-bindings for Alpao DMs
+and Python 3.8-binding for Alpao DMs are compatible with Linux systems, and other Python-bindings for Alpao DMs in Linux
 should also be compatible following their release. However, we note that SenAOReFoc is a control software, and the 
 performance of the closed-loop AO correction in practice is inevitably dependent on the state of the optical system.
 Finally, SenAOReFoc takes care of some reported issues in the field, such as the thermal effects of electromagnetic DMs 
