### 2022-03-04 #4075 @jiahecui
@diehlp Hello and sorry for waiting, we have now addressed the final item in response to Reviewer 1 and provided concluding remarks for our tests on different operating systems. These can now be found in lines 68-73 of the paper, and are also attached below for convenience.

"_The software fully operates in Windows under both ‘debug mode’ and ‘standard mode’ for hardware control. For Linux and macOS, functionality tests in ‘debug mode’ passes for both systems when hardware packages are not imported. Python-bindings for Ximea cameras and Python 3.8-binding for Alpao DMs are compatible with Linux systems, and other Python-bindings for Alpao DMs in Linux should also be compatible following their release._"

Thanks again and please let us know if any further information is required.
