### 2022-03-05 #4075 @jiahecui
Hi @diehlp, thank you very much for the instructions!

> Do a new release including all changes during the review and post the version here.

The version of the new release is v1.0.0.

> Upload this release to Zenodo or FigShare and post the DOI here. Note that the title and authors have to match with the paper.

This release has been uploaded to Zenodo. The DOI is 10.5281/zenodo.6330145, I've checked the title and authors, but please let me know if you spot any other problems with the Zenodo archiving. Many thanks again!
