### 2022-03-07 #4075 @diehlpk
@openjournals/dev Could you please have a look into the issues with the recommend-accept command?
