### 2022-03-07 #4075 @arfon
@diehlpk – looks like there's an issue with the font?

> Warning:  Missing character: There is no ∼ (U+223C) (U+223C) in font [lmsans9-regular]:+tlig;!
