### 2022-03-07 #4075 @diehlpk
@arfon Yes, but why does `@editorialbot generate pdf` works and `@editorialbot recommend-accept ` not?
