### 2022-03-08 #4075 @diehlpk
@jiahecui I am so sorry, but there is another issue with the references
```

Run bundle exec ruby /home/runner/work/_actions/xuanxu/validate-xml-files-action/main/validate_files.rb
!! Invalid Crossref XML v5.3.1 in tmp/f63b45c927f61f1c9339c2c4251da6d2/paper/paper.crossref. Errors: 2.
  - 371:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet maxLength] The value has a length of 19; this exceeds the allowed maximum length of 9.
  - 371:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet pattern] The value 1538-9235 1040-5488 is not accepted by the pattern d{4}-?d{3}[dX].
```
