+++ paper/paper.bib
@@ -218,8 +218,6 @@
   volume  = {87},
   number  = {12},
   pages   = {930-941},
-  issn    = {1538-9235
-1040-5488},
   doi     = {10.1097/OPX.0b013e3181ff9a8b},
   type    = {Journal Article},
   url     = {https://pubmed.ncbi.nlm.nih.gov/21057346
