### 2022-03-10 b03cfd5
Update paper.md
Fixed some formatting in the JOSS paper
