### 2022-03-10 #4075 @kyleniemeyer
Hi @jiahecui, I submitted a small PR that fixes some formatting issues in the paper: https://github.com/jiahecui/SenAOReFoc/pull/1

Could you merge this? I can then accept the paper.
